qso ordering by band as number

+++ cfmrda/srv.py
@@ -1342,7 +1342,7 @@ support@cfmrda.ru"""
             qso.rda = %(rda)s and
             (qso.band = %(band)s or %(band)s is null) and
             (qso.mode = %(mode)s or %(mode)s is null)
-        order by qso.band, qso.mode
+        order by qso.band::numeric, qso.mode
         """,\
         'activator': """
             select json_build_object('mode', mode,
@@ -1364,7 +1364,7 @@ support@cfmrda.ru"""
                             (qso.mode = %(mode)s or %(mode)s is null)
                         group by qso.upload_id, user_cs, upload_type, 
                             mode, band, qso.rda, dt
-                        order by band, mode) as l_0
+                        order by band::numeric, mode) as l_0
         """}
 
         @asyncio.coroutine
